requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
python-multipart==0.0.12
anthropic==0.39.0
google-generativeai==0.8.3
gspread==6.1.4
google-auth==2.35.0
httpx==0.27.2
pydantic==2.9.2
openpyxl==3.1.5
